refactor(Interface): create interface-package
